Return image depth not channel depth
We want to return MagickGetImageDepth instead of
MagickGetImageChannelDepth. This was causing problems on PNG images that
used a single color on a transparent background. In most cases the
channel depth matched the image depth but in the PNG case the channel
depth would be 5 bits while the image depth would be 8 bits, causing the
images to not render correctly on the PDF.

Fixes #8